+++ src/tankrotationexample/game/Explosive.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public abstract class Explosive extends Bullet{
 
-    Explosive(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    Explosive(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
