+++ src/tankrotationexample/game/GodMode.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public class GodMode extends Powerup{
 
-    GodMode(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    GodMode(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
