+++ src/tankrotationexample/game/NonExplosive.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public abstract class NonExplosive extends Bullet{
 
-    NonExplosive(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    NonExplosive(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
