+++ src/tankrotationexample/game/RocketLauncherMode.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public class RocketLauncherMode extends Powerup{
 
-    RocketLauncherMode(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    RocketLauncherMode(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
